requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
hijri-converter>=2.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
